src/requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.2
django-debug-toolbar==1.9.1
django-extensions==2.0.2
django-filter==1.1.0
django-filters==0.2.1
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-gis==0.12
drf-extensions==0.3.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.1
openapi-codec==1.3.2
pandas==0.22.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
